whatspp/requirements.txt
flask
python-dotenv
requests
python-dotenv
